requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
reportlab==4.0.7
weasyprint==60.2
PyJWT==2.8.0
cryptography==41.0.8
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-filter==23.5
markdown==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0